accounts - cannot delete account if it has transactions or transfers

+++ app/Http/Controllers/Admin/Accounting/AccountsController.php
@@ -62,9 +62,12 @@ class AccountsController extends Controller
             return response()->json([], 422);
         }
 
-//        if (Account::where('category_id', $id)->count()) {
-//            return response()->json([], 409);
-//        }
+        if ($account->transactions()->count() ||
+            $account->toTransfers()->count() ||
+            $account->fromTransfers()->count())
+        {
+            return response()->json([], 409);
+        }
 
         $account->delete();
